# [Graphbolt] Enable Link Prediction example model.inference to accelerate the evaluation speed