## Description
<!-- Brief description. Refer to the related issues if existed.
It'll be great if relevant reviewers can be assigned as well.-->

Given that the original sampling-based evaluation function is considerably slower compared to the DGL inference-based evaluation, enabling the `model.inference` in the Link Prediction example is crucial for a more efficient and straightforward regression test.

### Performance
The updated evaluation function significantly reduces processing time, requiring less than 1 hour as opposed to over 12 hours with the original evaluation function.

Train log:
```
$ python examples/sampling/graphbolt/link_prediction.py --early-stop 1000 --train-batch-size 1024 --eval-batch-size 4096 --epochs 10 --device cuda --num-workers 4
Training in cuda mode.
Loading data
The dataset is already preprocessed.
Training...
  0%|                               | 0/10 [00:00<?, ?it/s]
Epoch 00000 | Loss 0.1416 | Time 97.7365 s
 10%|██▎                    | 1/10 [01:37<14:39, 97.74s/it]
Epoch 00001 | Loss 0.0880 | Time 101.1375 s
 20%|████▌                  | 2/10 [03:18<13:17, 99.74s/it]
Epoch 00002 | Loss 0.0799 | Time 98.7655 s
 30%|██████▉                | 3/10 [04:57<11:35, 99.29s/it]
Epoch 00003 | Loss 0.0738 | Time 100.2421 s
 40%|█████████▏             | 4/10 [06:37<09:58, 99.67s/it]
Epoch 00004 | Loss 0.0705 | Time 99.0592 s
 50%|███████████▌           | 5/10 [08:16<08:17, 99.45s/it]
Epoch 00005 | Loss 0.0679 | Time 99.2085 s
 60%|█████████████▊         | 6/10 [09:56<06:37, 99.37s/it]
Epoch 00006 | Loss 0.0657 | Time 99.6363 s
 70%|████████████████       | 7/10 [11:35<04:58, 99.46s/it]
Epoch 00007 | Loss 0.0633 | Time 101.9516 s
 80%|█████████████████▌    | 8/10 [13:17<03:20, 100.25s/it]
Epoch 00008 | Loss 0.0618 | Time 98.2041 s
 90%|████████████████████▋  | 9/10 [14:55<01:39, 99.61s/it]
Epoch 00009 | Loss 0.0606 | Time 98.8265 s
100%|██████████████████████| 10/10 [16:34<00:00, 99.48s/it]
Testing...
Start node embedding inference.
715it [00:02, 274.28it/s]
715it [00:02, 267.21it/s]
715it [00:02, 279.65it/s]
Evaluate: 100%|████████████| 22/22 [00:16<00:00,  1.32it/s]
Evaluate: 100%|████████████| 22/22 [00:16<00:00,  1.35it/s]
Validation MRR 0.7480, Test MRR 0.7492
```

## Checklist
Please feel free to remove inapplicable items for your PR.
- [x] The PR title starts with [$CATEGORY] (such as [NN], [Model], [Doc], [Feature]])
- [x] I've leverage the [tools](https://docs.google.com/document/d/1iHyj7zlmygKSk5gBPsqIqL5ASPzJSPREaNT_QdsiYA4/edit) to beautify the python and c++ code.
- [x] The PR is complete and small, read the [Google eng practice (CL equals to PR)](https://google.github.io/eng-practices/review/developer/small-cls.html) to understand more about small PR. In DGL, we consider PRs with less than 200 lines of core code change are small (example, test and documentation could be exempted).
- [ ] All changes have test coverage
- [x] Code is well-documented
- [x] To the best of my knowledge, examples are either not affected by this change, or have been fixed to be compatible with this change
- [x] Related issue is referred in this PR
- [x] If the PR is for a new model/paper, I've updated the example index [here](../examples/README.md).

## Changes
<!-- You could use following template
- [ ] Feature1, tests, (and when applicable, API doc)
- [ ] Feature2, tests, (and when applicable, API doc)
-->